Thread-safe handler, better event loop
This commit adds a RWMutex into the handler code to make it thread-safe.
It also added a proper contextful timer for graceful cancellation and
cleanup, as well as minor modifications to potentially speed things up.

This commit also fixes a compile error in package api, presumably caused
by a prior mistake.