feat!: Add RoomMessageEvent.InReplyTo
This commit changes the existing RelatesTo field's type to a
json.RawMessage, as the latest specs now describe it as an extensible
object.

A new method named InReplyTo is added to parse the m.in_reply_to object
inside the RelatesTo object. Its purpose is to allow clients to see
which event the message is replying to.