feat(encrypt)!: Add encrypt.Client
This commit adds encrypt.Client and wraps the /keys/X endpoints. It also
adds the needed types and helper methods for those types. Some types
that are created in this commit will be used outside of the Client as
well.

This commit also removes some encrypt-specific fields from
api.SyncResponse in favor of encrypt's own SyncResponse, which the user
can use by calling encrypt.ParseSyncResponse with the api.SyncResponse.
This is done to prevent possibly cyclical imports while ensuring that
the right types are used.

In order to do that, api.SyncResponse now keeps the raw JSON of every
single field. This not only allows encrypt to do what it's doing, but it
also allows the user to implement whatever API that extends its /sync.
It may be a bit too wasteful, but future considerations can be made to
remove already-parsed fields.